Restructure drm blueprint to include drm specific templates, protocol code, and assets
