+++ library_registry/admin/routes.py
@@ -1,20 +1,8 @@
-from flask import Blueprint, render_template, request, Response, current_app, session, redirect, url_for
+from flask import Blueprint, current_app
 
 from library_registry.decorators import (
-    compressible,
-    has_library,
     returns_json_or_response_or_problem_detail,
     returns_problem_detail,
-    uses_location,
-)
-
-from library_registry.model import (
-    Admin,
-    production_session,
-)
-
-from library_registry.problem_details import (
-    INVALID_CREDENTIALS,
 )
 
 admin = Blueprint('admin', __name__)
