+++ library_registry/drm/adobe_vendor_id.py
@@ -3,7 +3,7 @@ import re
 import requests
 from flask import Response, request
 
-import library_registry.adobe_xml_templates as t
+import library_registry.drm.templates.adobe_xml_templates as t
 from library_registry.util.short_client_token import ShortClientTokenDecoder
 from library_registry.util.string_helpers import base64
 from library_registry.util.xmlparser import XMLParser
