+++ library_registry/library_protocol/__init__.py
@@ -1,1 +1,1 @@
-from .library_protocol_blueprint import libr
+from .routes import libr
